+++ cards/admin.py
@@ -1,6 +1,9 @@
 from django.contrib import admin
-from .models import CardTemplate
+from .models import CardTemplate, CardInstance
 
 @admin.register(CardTemplate)
 class CardTemplateAdmin(admin.ModelAdmin):
-    list_display = ('name', 'creator', 'created_at')
+    list_display = ('name', 'creator', 'is_public')
+    search_fields = ('name',)
+
+admin.site.register(CardInstance)